chore: refactor to notes, rm hero
